# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Band Manager API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /me:
    get:
      operationId: getCurrentAccount
      summary: Get the authenticated user and active band
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current account context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentAccountResponse'
        '401':
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Authenticated user has not completed onboarding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    CurrentAccountResponse:
      type: object
      required:
        - user
        - activeBand
      properties:
        user:
          $ref: '#/components/schemas/User'
        activeBand:
          $ref: '#/components/schemas/BandMembership'
    User:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
    BandMembership:
      type: object
      required:
        - bandId
        - bandName
        - role
        - canWrite
      properties:
        bandId:
          type: string
          format: uuid
        bandName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        canWrite:
          type: boolean
...
    Role:
      type: string
      enum:
        - owner
        - admin
        - member
        - viewer
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string